Faulds & battleskirts, new clothing! (#1804)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds faulds and battleskirts! More clothing for people to choose from,
they're colorable with the dye bin and can be worn on the pants, cloak,
or belt slot.
Colored Faulds:

![image](https://github.com/user-attachments/assets/67cc357c-1fbe-496f-aafe-c02687f8aa16)
Battleskirts:

![image](https://github.com/user-attachments/assets/2ec9f862-b177-4e5f-b0eb-5c72d531661a)
Faulds:

![image](https://github.com/user-attachments/assets/8c792149-953f-40e8-8b3b-05fa9a1539fb)

Also credits to this guy again:

![image](https://github.com/user-attachments/assets/437df01f-7faa-4490-8f76-fc403dfde672)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More choices for clothing! More diversity and such.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
# Conflicts:
#	code/modules/clothing/rogueclothes/cloaks.dm
#	code/modules/roguetown/roguecrafting/sewing.dm
#	code/modules/roguetown/roguejobs/tailor/dyer.dm
#	icons/roguetown/clothing/cloaks.dmi
#	icons/roguetown/clothing/onmob/cloaks.dmi