smallish i hope buff and changes to acrhfey warlock
changes archfey warlock to use it's own version of seelie kiss and strip with a stronger kiss and lower cooldowns

the reason for the buffs? i just think it would be funny to lower the CD of strip but didn't want to do it for every damm seelie, the kiss i am pretty sure it only injects 1u of health and mana pot so it's kinda useless but i don't know how good it would be with 12u of both, probably too much for mana?

also change their spell list, might revert the buff to strip and the revomal of animate object

i don't expect this to actually get approved because i don't even know if it works or the whole balance thing! also i am scared of archfey patron seelies with dupes of their own spells... if i knew what to do to prevent that i would but i am no coder (i also have no clue how git works)