Mining virtue nerf
We're going in a direction where people are supposed to have to talk to others, not just do everything themselves with a free steel pick and good mining  skills. Alchemists are gamin a little too close to the sun (Myself included. But I think it's for the best). Hire a miner, interact, stimulate the economy.